shifted the gender code to new file
